Handled JSON loading, set up scaffolding for http request for table, and corrected conflicting css files
It is set up now to take in a JSON and load the information into a table through various functions. I tried to add indepth comments to most of the code we already had and any new code. When we get to the point to load in the table data from an endpoint we will just change one location