Quick Environment setup/fix for backend
allowed addition of .env and enforced its use. This is universal for all of us so I included it for now, but I will edit the .gitignore to remove it and the package.json and such later. For now I included them so we can all have it work seemlessly